pyproject.toml
[tool.poetry]
name = "auto-todo"
version = "0.1.0"
description = ""
authors = ["MaxAFriedrich <58374387+MaxAFriedrich@users.noreply.github.com>"]
packages = [{include = "auto_todo"}]

[tool.poetry.dependencies]
python = "^3.11"
pyyaml = "^6.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
